Rewrote functional options as methods on a singleton struct instance.
This makes the options more discoverable via smart completion provided
by most text editors and IDEs. We get all the benefits of both the
fluent builder pattern and the functional options pattern with none of
their disadvantages:

	scanner := csv.NewScanner(
		strings.NewReader(in),
		csv.Options.Comma(';'),
		csv.Options.Comment('#'),
	)

The `options` type is not exported because the caller need not create
new instances.